# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super TTS daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/tts/super-tts-http.sock` (override with `SUPER_TTS_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/tts/super-tts-http.sock\" \\\n     -H \"Authorization: Bearer $TTS_TOKEN\" \\\n     http://tts.local/v1/ping\n```",
    "contact": {
      "name": "Super TTS",
      "url": "https://github.com/jorge-menjivar/super-tts"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://tts.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/backend/list": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "List installed backends",
        "description": "Every backend installed on this machine, each with the models it serves, the options it exposes, and which of its secrets are configured. Secret *values* are never returned — only whether each is set.\n\nThis is the full catalog. `GET /pipeline/{stage}/model/list` is the narrower read a stage's model picker wants, and browsing what is *available to install* is `GET /registry/backend/list`.",
        "operationId": "list_backends",
        "responses": {
          "200": {
            "description": "The installed catalog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendCatalog"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/backend/{backend_id}/secret/{name}": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "Check whether one secret is set",
        "description": "Reports existence only. There is no endpoint that returns a stored credential, so a client that has forgotten what it wrote can only overwrite it.",
        "operationId": "get_backend_secret",
        "parameters": [
          {
            "name": "backend_id",
            "in": "path",
            "description": "The backend's id — its `source` as `GET /backend/list` reports it — percent-encoded, e.g. `github.com%2Fsuper-tts%2Fopenai`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Fsuper-tts%2Fopenai"
          },
          {
            "name": "name",
            "in": "path",
            "description": "The secret's name, as the backend's manifest declares it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Whether a value is stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretConfigured"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `secrets` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such secret (`unknown_secret`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "secrets"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "backends"
        ],
        "summary": "Store a secret",
        "description": "Writes the credential to the system keyring, replacing whatever was there. It cannot be read back afterwards; only replaced or cleared.\n\nIf the stage is currently running a model from this backend, the daemon reloads it so the new credential is used at once — a key rotated while a model is loaded takes effect without the user knowing a reload was needed.",
        "operationId": "set_backend_secret",
        "parameters": [
          {
            "name": "backend_id",
            "in": "path",
            "description": "The backend's id — its `source` as `GET /backend/list` reports it — percent-encoded, e.g. `github.com%2Fsuper-tts%2Fopenai`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Fsuper-tts%2Fopenai"
          },
          {
            "name": "name",
            "in": "path",
            "description": "The secret's name, as the backend's manifest declares it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretConfigured"
                }
              }
            }
          },
          "400": {
            "description": "The value was empty (`invalid_request`). Use `DELETE` to clear a secret.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `secrets` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such secret (`unknown_secret`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The keyring could not be written (`keyring_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "secrets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "backends"
        ],
        "summary": "Clear a secret",
        "description": "Removes the stored credential from the keyring, returning the secret to unset. A backend that requires it will refuse to load until one is stored again.\n\nThis is the operation to reach for when a user disconnects an account: the value is gone from the keyring, not merely ignored.",
        "operationId": "delete_backend_secret",
        "parameters": [
          {
            "name": "backend_id",
            "in": "path",
            "description": "The backend's id — its `source` as `GET /backend/list` reports it — percent-encoded, e.g. `github.com%2Fsuper-tts%2Fopenai`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Fsuper-tts%2Fopenai"
          },
          {
            "name": "name",
            "in": "path",
            "description": "The secret's name, as the backend's manifest declares it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretConfigured"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `secrets` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such secret (`unknown_secret`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The keyring could not be written (`keyring_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "secrets"
            ]
          }
        ]
      }
    },
    "/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Subscribe to the event stream",
        "description": "A Server-Sent Events stream of everything the daemon publishes. The response is `text/event-stream` and stays open until the client disconnects, the daemon shuts down, or the calling binary changes on disk (which emits a final `revoked` frame and invalidates the token).\n\nThe first frame is always `subscribed`, carrying a `client_id` and the topic list that was accepted. A `: keepalive` comment is sent periodically so idle connections are not reaped by intermediaries. Subsequent frames set `event:` to the topic name and `data:` to that topic's payload, with no outer envelope.\n\n**Scopes are per topic, not per endpoint.** Any valid token reaches this route; the subscription is then refused whole if the token lacks the scope for *any* requested topic, so ask only for what you hold.\n\n| Topic | Scope | Payload |\n|---|---|---|\n| `speaking_state` | `playback_events` | `{ is_speaking, utterance_id? }` |\n| `speech_progress` | `playback_events` | `{ utterance_id, spoken_ms, queued_ms }` |\n| `frequency_bands` | `audio_visualization` | `{ bands_b64, sample_rate, total_energy }` |\n| `daemon_status_changed`, `download_progress`, `registry_install` | `daemon_status` | see `docs/protocol/endpoints/v1/events.md` |\n\nThis is how a client follows an utterance past the request that started it: `POST /speak` returns as soon as the audio is queued, and `speaking_state` / `speech_progress` outlive the connection that asked for it.\n\nUnder backpressure the daemon drops frames rather than buffering them without bound; `frequency_bands` is what sheds in practice, since it is emitted many times per second and the control frames are not.",
        "operationId": "events",
        "parameters": [
          {
            "name": "topics",
            "in": "query",
            "description": "Comma-separated topic names, at least one. Unknown or empty → `400 invalid_topic`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "speaking_state,speech_progress"
          }
        ],
        "responses": {
          "200": {
            "description": "The stream is open. Frames follow as `event:`/`data:` pairs.",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "`topics` was missing, empty, or named a topic that does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope for one of the requested topics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackendCatalog": {
        "type": "object",
        "description": "Every installed backend, with its models, options and secrets.",
        "required": [
          "status",
          "backends"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackendInfo"
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "SecretConfigured": {
        "type": "object",
        "description": "Whether one secret is set.",
        "required": [
          "status",
          "configured"
        ],
        "properties": {
          "configured": {
            "type": "boolean",
            "description": "Whether a value is stored."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Absent on write and clear, which answer about the secret just addressed."
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}